// src/types/Beatmap.ts
export interface RawBeatmap {
	beatmapset_id: string;
	beatmap_id: string;
	approved: string;
	total_length: string;
	hit_length: string;
	version: string;
	file_md5: string;
	diff_size: string;
	diff_overall: string;
	diff_approach: string;
	diff_drain: string;
	mode: string;
	count_normal: string;
	count_slider: string;
	count_spinner: string;
	submit_date: string;
	approved_date: string;
	last_update: string;
	artist: string;
	artist_unicode: string;
	title: string;
	title_unicode: string;
	creator: string;
	creator_id: string;
	bpm: string;
	source: string;
	tags: string;
	genre_id: string;
	language_id: string;
	favourite_count: string;
	rating: string;
	storyboard: string;
	video: string;
	download_unavailable: string;
	audio_unavailable: string;
	playcount: string;
	passcount: string;
	packs: string;
	max_combo: string;
	diff_aim: string;
	diff_speed: string;
	difficultyrating: string;
}

export interface Beatmap {
	beatmapset_id: number;
	beatmap_id: number;
	approved: 0 | 1 | 2 | 3 | 4 | number;
	total_length: number;
	hit_length: number;
	version: string;
	file_md5: string;
	diff_size: number;
	diff_overall: number;
	diff_approach: number;
	diff_drain: number;
	mode: 0 | 1 | 2 | 3 | number;
	count_normal: number;
	count_slider: number;
	count_spinner: number;
	submit_date: Date;
	approved_date: Date;
	last_update: Date;
	artist: string;
	artist_unicode: string;
	title: string;
	title_unicode: string;
	creator: string;
	creator_id: number;
	bpm: number;
	source: string;
	tags: string;
	genre_id: number;
	language_id: number;
	favourite_count: number;
	rating: number;
	storyboard: boolean;
	video: boolean;
	download_unavailable: boolean;
	audio_unavailable: boolean;
	playcount: number;
	passcount: number;
	packs: string | null;
	max_combo: number;
	diff_aim: number;
	diff_speed: number;
	difficultyrating: number;
}